Rename 'instructions.js' to 'variables.js'
